bar chart metrics for SVM & decision tree, line chart metrics for neural networks visualization of svm & decision tree run directly within tool import csv

+++ src/main/java/org/mithras/machinelearning/decisiontree/DecisionTreeClassifier.java
@@ -5,8 +5,8 @@ public class DecisionTreeClassifier extends Tree
     public String criterion = "gini";
     public String splitter = "best";
     public int max_depth;
-    public float min_samples_split = 2;
-    public float min_samples_leaf = 1;
+    public int min_samples_split = 2;
+    public int min_samples_leaf = 1;
     public float min_weight_fraction_leaf = 0.0f;
     public int max_features;
     public int random_state;
